@@ -20,7 +20,7 @@
 package org.comixwall.pffw;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import org.json.JSONObject;
 
