@@ -24,7 +24,7 @@ import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Binder;
 import android.os.IBinder;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import com.jcraft.jsch.Channel;
 import com.jcraft.jsch.ChannelExec;
