@@ -19,7 +19,7 @@
 
 package org.comixwall.pffw;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
